+++ PyAPIReference/extra.py
@@ -1,4 +1,6 @@
-from PyQt5.QtWidgets import QAction
+import os
+from importlib.util import spec_from_file_location, module_from_spec
+from PyQt5.QtWidgets import QAction, QDialog, QLabel, QVBoxLayout
 
 def create_qaction(menu, text: str, shortcut: str="", callback: callable=lambda: print("No callback"), parent=None) -> QAction:
 	"""This function will create a QAction and return it"""
@@ -12,3 +14,13 @@ def create_qaction(menu, text: str, shortcut: str="", callback: callable=lambda:
 	action.triggered.connect(callback) # Connect callback to callback argument
 
 	return action # Return QAction
+
+def get_module_from_path(path: str):
+	filename = os.path.basename(path) # filename means only the filename without the path, e.g.: PyAPIReference/PyAPIReference/main.py -> main.py
+	filename_without_extension = os.path.splitext(filename)[0]
+
+	spec = spec_from_file_location(filename_without_extension, path)
+	module = module_from_spec(spec)
+	spec.loader.exec_module(module)
+
+	return module